refactor: RequestListContainer 비즈니스로직 정리
